move time tests into their own file
